gemini-live-ephemeral-tokens-websocket/functions/requirements.txt
firebase-functions
google-genai